interconnect: add clk-based icc provider support
For some devices it is useful to export clocks as interconnect providers,
if the clock corresponds to bus bandwidth.

For example, on MSM8996 the cluster interconnect clock should be
scaled according to the cluster frequencies. Exporting it as an
interconnect allows one to properly describe this as the cluster
bandwidth requirements.

Change-Id: I2efb4b2ee3d18984aba2d7311bb46668a58beb94
Tested-by: Yassine Oudjana <y.oudjana@protonmail.com>
Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Link: https://lore.kernel.org/r/20230512001334.2983048-3-dmitry.baryshkov@linaro.org
Signed-off-by: Georgi Djakov <djakov@kernel.org>
Git-Commit: 0ac2a08f42ce5c06d5d1216eac59c046961acd4f
Git-Repo: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
[afilipov@qti.qualcomm.com: Trivial change to maintain
 compliance with interconnect framework APIs in kernel 5.15]
Signed-off-by: Atanas Filipov <afilipov@qti.qualcomm.com>