USB: Fix qusb_phy_disable_phy() call in qusb_phy_port_state_work()
When a USB 2.0 hub is used, the qusb_phy_disable_phy() function
is called causing the USB PHY to be disabled and the clock
and phy regulators to be turned off. This results in a failure to
enumerate the USB device.

Fix this by removing the call to qusb_phy_disable_phy() in the
qusb_phy_port_state_work() function during charger detection for
all charger detection except DCP as it only needs charging support.

Change-Id: I1867df138b965399d9a0acd733e5b6f384ae0720
Signed-off-by: Akash Kumar <akakum@qti.qualcomm.com>