i2c: i2c-msm-geni: Update i2c frequency table to match hardware guidance
With the current settings, the I2C buses are achieving around 370KHz
instead of the expected 400KHz. For 100KHz and 1MHz, the settings are
now more compliant and adhere to the Qualcomm’s internal programming
guide.

Update the I2C frequency table to align with the recommended values
outlined in the I2C hardware programming guide, ensuring proper
communication and performance.

Change-Id: Ie14a40eee846c0ea29bec512d6320e9548c508a4
Signed-off-by: Panicker Harish <quic_pharish@quicinc.com>